pyproject.toml
[build-system]
requires = ["setuptools>=72", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name            = "cropguard-ai"
version         = "4.0.0"
description     = "Plant disease detection using Vision Transformers"
readme          = "README.md"
requires-python = ">=3.10"
license         = { text = "MIT" }
authors         = [{ name = "CropGuard AI Team" }]

keywords = ["plant-disease", "deep-learning", "vision-transformer", "streamlit"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "streamlit>=1.35",
    "torch>=2.1",
    "torchvision>=0.16",
    "timm>=0.9",
    "Pillow>=10",
    "pandas>=2.0",
    "plotly>=5.18",
    "python-dotenv>=1.0",
    "fpdf2>=2.7",
]

[project.optional-dependencies]
train = [
    "torchmetrics>=1.3",
    "tqdm>=4.66",
    "gdown>=5.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "black>=24",
    "ruff>=0.4",
]
gradcam = [
    "grad-cam>=1.5",
]

[project.scripts]
cropguard-train = "train:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "pages*", "utils*", "data*", "models*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"
